feat(auth): Renamed file useLogout -> useLogoutUser
